refactor(#1017): share Position serializers between combat + areas
Move PositionSummarySerializer and PositionAdjacencyItemSerializer from
world/combat/serializers.py into the new world/areas/positioning/serializers.py
module so the upcoming non-combat scene serializer can import them without
duplication. Combat re-imports both names — no behavior change.
